scripts/requirements-drive-watcher.txt
google-api-python-client>=2.100.0
google-auth>=2.23.0
httpx>=0.25.0
